Created new class for new plugin menu
